# pay-guard-backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/branches/{branchId}/users": {
      "post": {
        "operationId": "Users_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "Users_list",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/branches/{branchId}/users/{userId}/remove": {
      "post": {
        "operationId": "Users_remove",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/platform/provider-incidents": {
      "get": {
        "operationId": "VerifyEtIncident_list",
        "parameters": [
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "HIGH",
                "CRITICAL"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "ACKNOWLEDGED"
              ]
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 10000,
              "default": 0,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Platform Provider Incidents"
        ]
      }
    },
    "/api/v1/platform/provider-incidents/{incidentId}": {
      "get": {
        "operationId": "VerifyEtIncident_require",
        "parameters": [
          {
            "name": "incidentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Platform Provider Incidents"
        ]
      }
    },
    "/api/v1/platform/provider-incidents/{incidentId}/acknowledge": {
      "post": {
        "operationId": "VerifyEtIncident_acknowledge",
        "parameters": [
          {
            "name": "incidentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeVerifyEtIncidentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Platform Provider Incidents"
        ]
      }
    },
...
  },
  "info": {
    "title": "PayGuard API",
    "description": "Canonical API contract for the PayGuard web portal, Android app, workers, and third-party integrations.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned by POST /api/v1/auth/login."
      }
    },
    "schemas": {
      "LoginContextDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "membershipRoleId": {
            "type": "string",
            "format": "uuid"
          },
          "workAssignmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "membershipId",
          "membershipRoleId"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identity": {
            "type": "string",
            "minLength": 3,
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "deviceIdentifier": {
            "type": "string",
            "minLength": 16,
            "maxLength": 512
          },
          "devicePlatform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web"
            ]
          },
          "context": {
            "$ref": "#/components/schemas/LoginContextDto"
          }
        },
        "required": [
          "identity",
          "password"
        ]
      },
...
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "temporaryPassword": {
            "type": "string",
            "minLength": 12
          },
          "role": {
            "type": "string",
            "enum": [
              "MANAGER",
              "CASHIER",
              "WAITER"
            ]
          },
          "address": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "Male",
              "Female"
            ]
          }
        },
        "required": [
          "email",
          "temporaryPassword",
          "role"
        ]
      },
      "RemoveStaffDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 5
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "AcknowledgeVerifyEtIncidentDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}